dom/apps/tests/test_apps_service.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=754141
-->
<window title="Mozilla Bug 754141"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=754141"
     target="_blank">Mozilla Bug 754141</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 754141 **/

  var appsService = Components.classes['@mozilla.org/AppsService;1']
                              .getService(Components.interfaces.nsIAppsService);
  SimpleTest.ok(appsService, "Should be able to get the Apps Service");

  SimpleTest.ok('getAppByManifestURL' in appsService,
                "getAppByManifestURL() should be a method in nsIAppsService");

  SimpleTest.is(appsService.getAppByManifestURL(''), null,
                "getAppByManifestURL() should return null for an empty string manifest url");

  ]]>
  </script>
</window>
